Crate gatk_haplotypecaller

GATK-RS HaplotypeCaller Library This crate contains the HaplotypeCaller implementation for GATK-RS.

Public API surface

Default features expose the HC engine / product API used by gatk-cli (run/engine/region/assembly/genotyping/PairHMM/VCF emit). Internal parity TSV exporters require the dev-dumps feature (off by default; enabled via gatk-cli and parity_harness). P12 fixture tables and discovery/compat helpers are crate-private unless parity_harness is enabled (integration tests + the hc_full_parity_gate example).

Re-exports

pub use compatibility::java_hc_site_semantics;

pub use active_region::TraversalTile;

pub use active_region::DEFAULT_TRAVERSAL_TILE_BP;

pub use activity_profile::activity_profile_base_process_state;

pub use activity_profile::adaptive_filter_size;

pub use activity_profile::band_pass_process_state;

pub use activity_profile::make_gaussian_kernel;

pub use activity_profile::normal_distribution;

pub use activity_profile::ActivityEvidence;

pub use activity_profile::ActivityProfileRegion;

pub use activity_profile::ActivityProfileState;

pub use activity_profile::ActivityProfileStateKind;

pub use activity_profile::BandPassActivityProfile;

pub use activity_profile::BandPassActivityProfileParams;

pub use activity_profile::PositiveSigma;

pub use activity_profile::GATK_BAND_PASS_DEFAULT_SIGMA;

pub use activity_profile::GATK_BAND_PASS_MAX_FILTER_SIZE;

pub use activity_profile::GATK_BAND_PASS_MIN_PROB_TO_KEEP_IN_FILTER;

pub use activity_profile::GATK_DEFAULT_ACTIVE_PROB_THRESHOLD;

pub use activity_profile::GATK_DEFAULT_MAX_PROB_PROPAGATION_DISTANCE;

pub use activity_scoring::approximate_log10_sum_log10_pair;

pub use activity_scoring::calc_ref_vs_any_log10_genotype_likelihoods;

pub use activity_scoring::calculate_multi_sample_any_non_ref_posterior;

pub use activity_scoring::calculate_single_sample_biallelic_non_ref_posterior;

pub use activity_scoring::haplotype_caller_activity_profile_state_multi_sample;

pub use activity_scoring::haplotype_caller_activity_profile_state_single_sample;

pub use activity_scoring::is_alt_after_assembly;

pub use activity_scoring::is_alt_before_assembly;

pub use activity_scoring::log10_sum_log10;

pub use activity_scoring::log_binomial_coefficient_natural;

pub use activity_scoring::normalize_from_log10_to_linear_space;

pub use activity_scoring::HaplotypeCallerActivityScoringParams;

pub use activity_scoring::PileupObservation;

pub use activity_scoring::AVERAGE_HQ_SOFTCLIPS_HQ_BASES_THRESHOLD;

pub use activity_scoring::LOG10_ONE_THIRD;

pub use activity_scoring::REF_MODEL_DELETION_QUAL;

pub use alignment::calculate_haplotype_cigar;

pub use alignment::Cigar;

pub use alignment::CigarOperator;

pub use alignment::SwParameters;

pub use allele_filter_options::ActiveRegionSpan;

pub use allele_filter_options::AlleleFilterOptions;

pub use allele_filtering::filter_assembly_and_likelihoods;

pub use allele_filtering::MAX_NON_REF_HAPLOTYPES_FOR_GENOTYPING;

pub use assembly::AssemblyGraphPruningParams;

pub use assembly::AssemblyGraphSummary;

pub use assembly_based_caller::assemble_reads;

pub use assembly_based_caller::assemble_reads_with_finalized;

pub use assembly_based_caller::call_region_assemble;

pub use assembly_based_caller::AssembleReadsArgs;

pub use assembly_based_caller::AssembledRegion;

pub use assembly_pipeline_stages::CallRegionAssemblyStage;

pub use assembly_pipeline_stages::EVENT_MAP_SYNC_AROUND_FILTER_RATIONALE;

pub use assembly_region_evaluator::add_locus_for_smoothed_activity;

pub use assembly_region_evaluator::evaluate_hc_activity_state;

pub use assembly_region_iterator::load_all_records_for_contig;

pub use assembly_region_iterator::load_all_records_for_contig_raw;

pub use assembly_region_iterator::load_records_for_shard_raw;

pub use assembly_region_iterator::refuse_oversized_assembly_region_reads;

pub use assembly_region_iterator::sync_read_qnames;

pub use assembly_region_iterator::AssemblyRegion;

pub use assembly_region_iterator::AssemblyRegionIterator;

pub use assembly_region_iterator::AssemblyRegionIteratorConfig;

pub use assembly_region_iterator::MAX_READS_PER_ASSEMBLY_REGION;

pub use assembly_region_trimmer::load_trim_variants_tsv;

pub use assembly_region_trimmer::trim_assembly_region;

pub use assembly_region_trimmer::AssemblyRegionTrimResult;

pub use assembly_region_trimmer::AssemblyRegionTrimmer;

pub use assembly_region_trimmer::AssemblyRegionTrimmerConfig;

pub use assembly_region_trimmer::TrimVariant;

pub use assembly_regions_dump::dump_smoothed_activity_tsv;

pub use assembly_result_set::AssemblyResultSet;

pub use bio_ids::AlleleDepth;

pub use bio_ids::AlleleIndex;

pub use bio_ids::DiploidGenotypeIndex;

pub use bio_ids::GenotypeQuality;

pub use bio_ids::HaplotypeIndex;

pub use bio_ids::KmerSize;

pub use bio_ids::PadOffset0;

pub use bio_ids::PhredLikelihood;

pub use bio_ids::Ploidy;

pub use bio_ids::ReadCoordinate;

pub use bio_ids::ReadDepth;

pub use bio_ids::ReadIndex;

pub use bio_ids::ReferenceCoordinate;

pub use bio_ids::SampleIndex;

pub use combine_gvcfs::run_combine_gvcfs;

pub use combine_gvcfs::CombineGvcfsArgs;

pub use engine::CallRegionArgs;

pub use engine::CallRegionMode;

pub use engine::CallRegionOutcome;

pub use engine::HaplotypeCallerEngine;

pub use event_map::Event;

pub use event_map::EventMap;

pub use event_map::IndelSpan;

pub use feature_context::FeatureContext;

pub use feature_context::FeatureDataSources;

pub use feature_context::FeatureLocatable;

pub use genome_loc::GenomeLoc;

pub use genome_loc::GenomePosition;

pub use genotype_gvcfs::run_genotype_gvcfs;

pub use genotype_gvcfs::GenotypeGvcfsArgs;

pub use genotype_gvcfs::DEFAULT_STAND_CALL_CONF;

pub use given_alleles::merge_given_alleles_into_assembly;

pub use given_alleles::GatkGivenAllele;

pub use gvcf_writer::gatk_hc_gvcf_header_lines;

pub use gvcf_writer::GvcfWriter;

pub use gvcf_writer::GvcfWriterConfig;

pub use gvcf_writer::GATK_HC_DEFAULT_GQB;

pub use haplotype::Haplotype;

pub use hc_genotyping_engine::biallelic_genotype_log10_likelihoods_gatk;

pub use hc_genotyping_engine::diagnose_genotype_variation_event;

pub use hc_genotyping_engine::genotype_active_region;

pub use hc_genotyping_engine::java_emit_af_decision;

pub use hc_genotyping_engine::java_vcf_shaped_rescue_gl;

pub use hc_genotyping_engine::marginalize_rows_to_biallelic_alleles;

pub use hc_genotyping_engine::subset_biallelic_haplotype_indices;

pub use hc_genotyping_engine::GenotypeRejectReason;

pub use hc_genotyping_engine::GenotypingSemantics;

pub use hc_genotyping_engine::HcGenotypingConfig;

pub use hc_genotyping_engine::InformativeAd;

pub use hc_genotyping_engine::JavaEmitAfDecision;

pub use hc_genotyping_engine::RegionGenotypeResult;

pub use hc_genotyping_engine::SparsePlShape;

pub use hc_genotyping_engine::DEFAULT_INFORMATIVE_READ_OVERLAP_MARGIN;

pub use hc_genotyping_engine::DEFAULT_STAND_EMIT_CONFIDENCE;

pub use hc_genotyping_engine::format_locus_genotype_pl_dump;

pub use hc_genotyping_engine::pairhmm_locus_trace_dump;

pub use hq_soft_clip::count_high_quality_soft_clip_bases_rcm;

pub use hq_soft_clip::hq_soft_clip_running_mean_at_locus;

pub use hq_soft_clip::max_hq_soft_clip_bases;

pub use hq_soft_clip::RCM_HQ_SOFT_CLIP_QUAL_THRESHOLD;

pub use junction_kbest::find_junction_best_haplotypes;

pub use junction_kbest::JunctionKBestPath;

pub use junction_tree_graph::build_junction_tree_graph_from_ref_and_reads;

pub use kbest_haplotype::find_best_haplotypes;

pub use kbest_haplotype::find_best_haplotypes_for_assembly;

pub use kbest_haplotype::find_best_haplotypes_preserving_cycles;

pub use kbest_haplotype::KBestPath;

pub use likelihood_engine::prepare_read_quals_for_pairhmm_inplace;

pub use likelihood_engine::score_read_against_haplotypes;

pub use likelihood_engine::HcLikelihoodEngineConfig;

pub use likelihood_engine::HcLikelihoodImplementation;

pub use locus_iterator::IntervalLocusIterator;

pub use locus_iterator::LocusPileupState;

pub use locus_iterator::LocusPileupWalker;

pub use minimal_genotyping::calculate_single_sample_ref_vs_any_active_state_profile_value;

pub use minimal_genotyping::haplotype_caller_activity_profile_state_minimal_genotyping;

pub use pairhmm::pairhmm_fp_eq;

pub use pairhmm::pairhmm_log10_likelihood;

pub use pairhmm::pairhmm_log10_likelihood_slices;

pub use pairhmm::pairhmm_log10_likelihoods_vectorized;

pub use pairhmm::pairhmm_log10_likelihoods_vectorized_slices;

pub use pairhmm::PairHmmFpPolicy;

pub use pairhmm::PairHmmInput;

pub use pairhmm::PairHmmParams;

pub use pairhmm_log10::log10_pairhmm_likelihood as log10_pairhmm_likelihood_exact;

pub use pairhmm_log10::log10_pairhmm_likelihood_parity_defaults;

pub use pairhmm_log10::GATK_PARITY_DEFAULT_DEL_QUAL;

pub use pairhmm_log10::GATK_PARITY_DEFAULT_GCP;

pub use pairhmm_log10::GATK_PARITY_DEFAULT_INS_QUAL;

pub use pairhmm_logless::logless_pairhmm_likelihood;

pub use pairhmm_logless::logless_pairhmm_likelihood_parity_defaults;

pub use pairhmm_logless::INITIAL_CONDITION;

pub use pairhmm_logless::INITIAL_CONDITION_LOG10;

pub use pairhmm_simd::best_simd_backend;

pub use pairhmm_simd::parse_pair_hmm_impl;

pub use pairhmm_simd::resolve_pair_hmm_impl;

pub use pairhmm_simd::score_read_haps_logless;

pub use pairhmm_simd::PairHmmBackend;

pub use pairhmm_simd::PairHmmImpl;

pub use read_downsample::apply_positional_downsampler;

pub use read_downsample::GatkJavaRng;

pub use read_downsample::PositionalDownsamplerConfig;

pub use read_downsample::GATK_DEFAULT_MAX_READS_PER_ALIGNMENT_START;

pub use read_downsample::GATK_JAVA_RANDOM_SEED;

pub use read_header_semantics::ReadHeaderSemantics;

pub use read_header_semantics::ResolvedReadHeaderSemantics;

pub use read_model::mapq_passes_minimum;

pub use read_model::passes_hc_read_filter_set;

pub use read_model::passes_hc_read_filters;

pub use read_model::passes_hc_read_filters_fields;

pub use read_model::passes_hc_read_filters_with_header;

pub use read_model::passes_printreads_parity_filters;

pub use read_model::standard_hc_read_filter_failure_index;

pub use read_model::HcReadFilterSet;

pub use read_model::ReadFilterParams;

pub use read_model::FLAG_DUPLICATE;

pub use read_model::FLAG_NOT_PRIMARY;

pub use read_model::FLAG_SEGMENT_UNMAPPED;

pub use read_model::FLAG_SUPPLEMENTARY;

pub use read_model::FLAG_VENDOR_QUALITY_FAILED;

pub use read_model::GATK_HC_DEFAULT_MIN_MAPPING_QUALITY;

pub use read_model::MAPPING_QUALITY_UNAVAILABLE;

pub use read_model::STANDARD_HC_READ_FILTER_JAVA_NAMES;

pub use read_projection::cigar_hard_clip_length;

pub use read_projection::cigar_soft_clip_ends;

pub use read_projection::query_index_at_reference_position;

pub use read_projection::reference_position_at_query_index;

pub use read_threading_assembler::assemble_from_ref_and_reads;

pub use read_threading_assembler::audit_threading_dangling_recovery;

pub use read_threading_assembler::AssemblyResult as ThreadingAssemblyResult;

pub use read_threading_assembler::AssemblyStatus;

pub use read_threading_assembler::ReadThreadingAssemblerArgs;

pub use read_threading_assembler::ThreadingDanglingAudit;

pub use read_threading_assembler::DEFAULT_NUM_BEST_HAPLOTYPES_PER_GRAPH;

pub use read_threading_assembler::audit_kbest_extract;

pub use read_threading_assembler::path_is_too_divergent_from_reference;

pub use read_threading_assembler::KbestExtractAuditRow;

pub use read_threading_assembler::KbestExtractReject;

pub use read_threading_assembler::probe_seq_graph_kmer_attempts;

pub use read_threading_assembler::SeqGraphKmerProbeRow;

pub use read_threading_graph::assembly_graph_from_ref_and_reads_threading_with_summary;

pub use read_threading_graph::reference_has_non_unique_kmers;

pub use read_threading_graph::threading_non_unique_summary;

pub use read_threading_graph::ThreadingNonUniqueSummary;

pub use read_transformer::apply_iupac_strict_transform;

pub use read_transformer::apply_shard_read_pipeline;

pub use read_transformer::load_contig_records_hc_production;

pub use read_transformer::ShardReadPipelineConfig;

pub use read_validation::validate_mapped_read_sanity;

pub use ref_confidence::calc_genotype_likelihoods_of_ref_vs_any;

pub use ref_confidence::reference_gq_from_log10_gl;

pub use ref_confidence::reference_model_for_no_variation_region;

pub use ref_confidence::InactiveReferenceModelOutcome;

pub use ref_confidence::ReferenceConfidenceConfig;

pub use ref_confidence::ReferenceConfidenceLocusDetail;

pub use reference_context::ReferenceContext;

pub use region_read_likelihood::RegionReadLikelihood;

pub use region_vcf_emit::try_emit_call_region_variant;

pub use region_vcf_emit::try_emit_call_region_variants;

pub use region_vcf_emit::HC_PIPELINE_ASSEMBLY_REGION_V1;

pub use region_vcf_emit::HC_PIPELINE_LEGACY_PROVISIONAL;

pub use region_vcf_emit::HC_PIPELINE_SCAFFOLD;

pub use run::run_haplotype_caller;

pub use shared_bam::empty_shared_record;

pub use shared_bam::empty_shared_record_ref;

pub use shared_bam::into_unique_records;

pub use shared_bam::is_empty_shared_record;

pub use shared_bam::record_make_mut;

pub use shared_bam::share_record;

pub use shared_bam::share_records;

pub use shared_bam::BamRecordSlot;

pub use shared_bam::SharedBamRecord;

pub use walker::make_read_shards;

pub use walker::make_read_shards_default_padding;

pub use walker::ReadShard;

pub use walker::GATK_DEFAULT_ASSEMBLY_REGION_PADDING;

pub use walker_apply::call_disposition;

pub use walker_apply::AssemblyRegionCallDisposition;

pub use walker_apply::WalkerApplyStats;

pub use walker_traversal::collect_assembly_regions;

pub use walker_traversal::drain_assembly_regions_for_shard;

pub use walker_traversal::flatten_assembly_regions;

pub use walker_traversal::for_each_assembly_region;

pub use walker_traversal::into_assembly_regions;

pub use walker_traversal::traverse_assembly_region_walker;

pub use walker_traversal::WalkerShardTraversal;

pub use walker_traversal::WalkerTraversal;

pub use walker_traversal::WalkerTraversalConfig;

pub use worker_threads::init_worker_threads;

pub use annotator_dump::dump_annotate_core_tsv;

pub use annotator_dump::dump_annotation_manifest_tsv;

pub use annotator_dump::dump_annotation_plugin_tsv;

pub use annotator_dump::dump_as_annotations_tsv;

pub use annotator_dump::dump_depth_per_sample_hc_tsv;

pub use annotator_dump::dump_excess_het_tsv;

pub use annotator_dump::dump_standard_annotations_tsv;

pub use assembler_args_dump::dump_assembler_args_tsv;

pub use assembly_debug_dump::dump_assembly_debug_stub_tsv;

pub use assembly_graph_dump::dump_assembly_assemble_tsv;

pub use assembly_graph_dump::dump_assembly_graph_dangling_summary_tsv;

pub use assembly_graph_dump::dump_assembly_graph_edges_tsv;

pub use assembly_graph_dump::dump_assembly_graph_low_quality_tsv;

pub use assembly_graph_dump::dump_assembly_graph_multi_kmer_edges_tsv;

pub use assembly_graph_dump::dump_assembly_graph_non_unique_summary_tsv;

pub use assembly_graph_dump::dump_assembly_graph_pruned_summary_tsv;

pub use assembly_graph_dump::dump_assembly_haplotype_cigars_tsv;

pub use assembly_graph_dump::dump_assembly_haplotypes_cap_tsv;

pub use assembly_graph_dump::dump_assembly_haplotypes_production_tsv;

pub use assembly_graph_dump::dump_assembly_haplotypes_tsv;

pub use assembly_graph_dump::dump_assembly_junction_haplotypes_tsv;

pub use assembly_graph_dump::dump_assembly_kbest_paths_tsv;

pub use assembly_graph_dump::dump_assembly_seqgraph_summary_tsv;

pub use assembly_graph_dump::dump_read_error_correction_tsv;

pub use assembly_graph_dump::load_assembly_reads_tsv;

pub use assembly_graph_dump::load_assembly_ref_tsv;

pub use assembly_graph_dump::write_assembly_graph_summary_tsv;

pub use assembly_graph_dump::write_dangling_recovery_summary_tsv;

pub use assembly_graph_dump::write_threading_non_unique_summary_tsv;

pub use assembly_region_assemble_dump::dump_assembly_region_haplotypes_tsv;

pub use assembly_region_assemble_dump::dump_assembly_region_kmer_probe_tsv;

pub use assembly_region_assemble_dump::AssemblyRegionHaplotypeTarget;

pub use assembly_region_finalize_dump::dump_assembly_region_assembly_stages_finalize_tsv;

pub use assembly_region_finalize_dump::dump_assembly_region_finalize_reads_tsv;

pub use assembly_region_genotype_dump::dump_assembly_region_genotype_subset_tsv;

pub use assembly_region_genotype_dump::dump_assembly_region_genotype_tsv;

pub use assembly_region_iterator::dump_assembly_region_features_tsv;

pub use assembly_region_iterator::dump_assembly_region_pileup_track_tsv;

pub use assembly_region_iterator::dump_assembly_region_reads_tsv;

pub use assembly_region_iterator::dump_assembly_region_reference_tsv;

pub use assembly_region_pairhmm_dump::dump_assembly_region_pairhmm_likelihoods_tsv;

pub use assembly_region_stages_dump::dump_assembly_region_assembly_stages_tsv;

pub use assembly_region_stages_dump::dump_assembly_region_kbest_paths_tsv;

pub use assembly_region_trimmer::dump_assembly_region_trim_tsv;

pub use assembly_regions_dump::dump_active_locus_tsv;

pub use assembly_regions_dump::dump_genotype_likelihood_activity_tsv;

pub use assembly_regions_dump::dump_locus_pileup_tsv;

pub use assembly_regions_dump::dump_locus_pileup_tsv_default_padding;

pub use assembly_regions_dump::dump_raw_activity_profile_tsv;

pub use assembly_regions_dump::dump_raw_activity_profile_tsv_with_contamination;

pub use assembly_regions_dump::dump_raw_activity_profile_tsv_with_force_calling;

pub use assembly_regions_dump::dump_smoothed_activity_profile_tsv;

pub use assembly_regions_dump::format_activity_prob;

pub use assembly_regions_dump::ForceCallingAllelesDump;

pub use assembly_regions_dump::GATK_HC_ALLELES_FEATURE_SOURCE;

pub use genotyping_dump::dump_af_em_tsv;

pub use genotyping_dump::dump_allele_subsetting_tsv;

pub use genotyping_dump::dump_force_calling_genotype_tsv;

pub use genotyping_dump::dump_genotype_format_tsv;

pub use genotyping_dump::dump_genotype_limits_tsv;

pub use genotyping_dump::dump_genotype_phasing_tsv;

pub use genotyping_dump::dump_genotyping_aggregate_tsv;

pub use genotyping_dump::dump_subset_alleles_integration_tsv;

pub use genotyping_dump::dump_subset_alleles_pl_tsv;

pub use genotyping_dump::dump_subset_alleles_vc_tsv;

pub use gvcf_writer_dump::dump_gvcf_header_tsv;

pub use gvcf_writer_dump::dump_gvcf_l5_merged_tsv;

pub use gvcf_writer_dump::dump_gvcf_writer_from_loci_fixture_tsv;

pub use hq_soft_clip_dump::dump_hq_soft_clip_mean_tsv;

pub use j_modes_dump::dump_dragen_mode_branch_tsv;

pub use j_modes_dump::dump_emit_mode_decision_tsv;

pub use likelihood_engine_dump::dump_likelihood_engine_config_tsv;

pub use likelihood_engine_dump::dump_likelihood_pcr_read_tsv;

pub use likelihood_engine_dump::dump_pcr_error_model_tsv;

pub use locus_pileup_detail_dump::dump_locus_pileup_detail_tsv;

pub use pairhmm_dump::dump_pairhmm_likelihoods_tsv;

pub use pairhmm_dump::load_pairhmm_cases_tsv;

pub use pairhmm_f3_dump::dump_pairhmm_bq_cap_tsv;

pub use pairhmm_f3_dump::dump_pairhmm_haplotype_filter_tsv;

pub use pairhmm_native_dump::dump_pairhmm_native_likelihoods_tsv;

pub use ploidy_dump::dump_ploidy_resolution_tsv;

pub use pre_dragstr_dump::dump_dragstr_calibration_tsv;

pub use read_downsample::dump_positional_downsample_summary_tsv;

pub use read_filter_dump::dump_hc_read_filter_tsv;

pub use read_filter_dump::HC_READ_FILTER_COUNT_SECTION;

pub use read_pre_len_dump::dump_read_pre_len_tsv;

pub use read_pre_mq_dump::dump_read_pre_mq_tsv;

pub use read_pre_overlap_dump::dump_read_pre_overlap_tsv;

pub use read_transformer_dump::dump_read_shard_pipeline_tsv;

pub use read_unclip_dump::dump_read_pre_softclip_tsv;

pub use ref_confidence_dump::dump_call_region_active_rcm_loci_tsv;

pub use ref_confidence_dump::dump_inactive_reference_model_tsv;

pub use ref_confidence_dump::dump_reference_confidence_locus_tsv;

pub use ref_confidence_merger_dump::dump_ref_confidence_merge_case;

pub use region_vcf_dump::dump_call_region_format_tsv;

pub use region_vcf_dump::dump_call_region_vcf_tsv;

pub use region_vcf_dump::dump_variant_format_from_gl_ad_tsv;

pub use region_vcf_dump::dump_variant_vcf_from_gl_ad_tsv;

pub use alignment::*;

pub use assembly::*;

pub use genotyping::*;

Modules

active_region

Fixed-width traversal tiles over closed genomic intervals. GATK’s adaptive active regions come later; this only splits each closed interval into non-overlapping tiles for engine wiring and logging.

activity_profile

GATK-aligned activity profile scaffolding (Phase 4, step 51). Mirrors the data model and numeric thresholds from GATK’s org.broadinstitute.hellbender.utils.activityprofile package and AssemblyRegionArgumentCollection defaults used by BandPassActivityProfile.

activity_scoring

Locus-by-locus activity scoring aligned with GATK HaplotypeCallerEngine#isActive. calc_ref_vs_any_log10_genotype_likelihoods mirrors ReferenceConfidenceModel#calcGenotypeLikelihoodsOfRefVsAny for the non–flow path (readsWereRealigned = false). This module ports the pre-band-pass probability path used for assembly-region discovery:

alignment

Haplotype-to-reference Smith–Waterman alignment (GATK CigarUtils.calculateCigar).

allele_filter_options

Typed options for crate::allele_filtering::filter_assembly_and_likelihoods (Sprint K-1 / R2).

allele_filtering

GATK AlleleFilteringHC.filterAlleles slice — drop weak haplotypes before genotyping. Java ranks haplotypes using per-allele getAlleleLikelihoodVsInverse (HC genotyping GL vs inverse allele). Rust uses that score when PairHMM rows exist, else read-LL sum fallback.

annotator_dump

Annotated variant site parity dumps.

assembler_args_dump

ReadThreadingAssemblerArgumentCollection defaults dump.

assembly

Local assembly graph scaffolding.

assembly_based_caller

GATK AssemblyBasedCallerUtils.assembleReads slice (finalize + padded ref + local assembly).

assembly_dangling_recovery

GATK AbstractReadThreadingGraph.recoverDanglingTails / recoverDanglingHeads parity.

assembly_debug_dump

E-D01 — assembly failure / graph debug dumps (scaffold).

assembly_graph_dump

Assembly graph edge dumps for L2 parity.

assembly_pipeline_stages

Typed assembly → genotyping pipeline stages (Sprint K-5). Documents the Java mutation sequence that call_region mirrors. The double [crate::read_event_discovery::sync_assembly_events_from_haplotype_cigars_with_harvest] call is intentional: EventMap must be current before filterAlleles and again after haplotype set changes.

assembly_pruning

GATK ChainPruner, LowWeightChainPruner, and AdaptiveChainPruner on crate::assembly::AssemblyGraph.

assembly_region_assemble_dump

Active assembly region → ReadThreadingAssembler haplotypes (GATK callRegion assembly slice).

assembly_region_evaluator

Per-locus HC activity before band-pass smoothing. Production path is free functions evaluate_hc_activity_state and add_locus_for_smoothed_activity (no evaluator trait / ZST shell).

assembly_region_finalize

GATK AssemblyBasedCallerUtils.finalizeRegion + assembly reference padding (GAP-E-06).

assembly_region_finalize_dump

ASM-1: per-read finalizeRegion parity (assembly-region-finalize-reads).

assembly_region_genotype_dump

Active region → assembly + genotyping slice (Java HcFullParityGateDump aligned).

assembly_region_iterator

GATK4 AssemblyRegionIterator–style region stream. See docs/ARCHITECTURE.md. Consumes a ReadShard padded span, walks 1-based loci with the same pileup → activity → band-pass path as the smoothed-activity parity exporters (feature dev-dumps), and drains BandPassActivityProfile::try_pop_next_ready_region like Java’s iterator (non-forcing pops after each locus, forcing flush at span boundaries). Shard load uses load_records_for_shard_raw + crate::read_transformer::apply_shard_read_pipeline in crate::walker_traversal::drain_assembly_regions_for_shard. load_all_records_for_contig still filters in one step for parity exporters that do not use the B.4 pipeline. B.3: after draining regions, use crate::walker_apply::WalkerApplyStats (or crate::HaplotypeCallerEngine::walker_apply_stats) to mirror Java apply count and inactive callRegion branching.

assembly_region_pairhmm_dump

E2E integration: active region → assembly → native Log10PairHMM read×haplotype matrix.

assembly_region_stages_dump

P12 ASM-1: staged assembly metrics (threading graph → SeqGraph cleanup).

assembly_region_trimmer

GATK AssemblyRegionTrimmer (GAP-B-06). Shrinks an assembly region to the span of assembled variation plus genotyping padding (post-assembly in Java HaplotypeCallerEngine.callRegion).

assembly_regions_dump

Dense smoothed activity profile export for parity with GATK HaplotypeCaller --assembly-region-out. Walks every 1-based position in -L, builds per-locus pileups, feeds BandPassActivityProfile, then writes position\\tactive_prob (smoothed) for comparison to the Java IGV derivative.

assembly_result_set

GATK AssemblyResultSet — multi-kmer assembly outcomes for one region (E2E.4).

bio_ids

Strongly typed biological identifiers for HaplotypeCaller. These newtypes exist to make domain mix-ups a compile error: haplotype index vs PL genotype index, 1-based genome locus vs pad-relative offset, MAPQ score vs SAM unavailable (255), etc. Prefer explicit constructors (new / try_new) over blanket From integer conversions. Convert at BAM/VCF wire edges only.

cigar

HTSJDK-style CIGAR elements for assembly haplotype alignment.

cigar_builder

GATK CigarBuilder parity (merge ops, trim end deletions).

combine_gvcfs

Multi-sample gVCF merge (GATK CombineGVCFs algorithm slice). CombineGVCFs — merge per-sample gVCFs into a multi-sample gVCF (no joint genotyping). Observable Java contract (GATK 4.4 CombineGVCFs): Union sample columns from input gVCFs. At each genomic breakpoint induced by record starts / END+1, merge overlapping reference-confidence VCs via ReferenceConfidenceVariantContextMerger semantics (crate::ref_confidence_merger::merge_reference_confidence). Remap PL/AD onto the unified allele list; missing samples emit no-call genotypes. Hom-ref <NON_REF> blocks carry INFO/END; variant sites emit discrete rows. This module is a Rust-native walker over that merger kernel — not a Java class clone.

compatibility

Java-compatibility / P12-interval semantics (Sprint I-1 quarantine). Java-compatibility / P12-interval semantics (waivers W-H1, W-H3). This module holds predicates and band tables derived from GATK 4.4 observable behavior on the NA12878 P12 validation slice (chr2:92300000–92350000). Not a claim of genome-wide GATK equivalence. Prefer phenotype predicates over new coordinate literals. See docs/ARCHITECTURE.md and docs/CLAIM_MATRIX.md. Sprint J: coupled_indel is the preferred recognition path for cluster indels; absolute band tables in java_hc_site_semantics are oracle/waiver surfaces.

emit_gates

Shared HC variant emit gates (Java AF threshold, read support, genotype emit parity). Leaf module breaking the hc_emit_policy ↔ hc_genotyping_engine cycle.

engine

HaplotypeCaller engine: traversal scaffold + callRegion assembly + likelihoods.

event_map

GATK Event / EventMap — genomic events implied by haplotype vs reference.

feature_context

GATK FeatureContext for assembly-region apply. Mirrors new FeatureContext(featureManager, assemblyRegion.getPaddedSpan). HC defaults use no backing resources (FeatureManager null); optional VCF paths supply forced / known alleles.

fragment_overlap

Overlapping paired-fragment base-quality correction. Mirrors AssemblyBasedCallerUtils.cleanOverlappingReadPairs and FragmentUtils.adjustQualsOfOverlappingPairedFragments.

genome_loc

1-based inclusive genomic interval (GATK SimpleInterval subset).

genotype_gvcfs

Joint genotyping of combined gVCFs (GATK GenotypeGVCFs algorithm slice). GenotypeGVCFs — joint genotyping of a multi-sample gVCF (post CombineGVCFs). Observable Java contract (GATK 4.4 GenotypeGVCFs / GenotypeGVCFsEngine): Input: one multi-sample gVCF with diploid PL/AD (typically CombineGVCFs output). Drop symbolic <NON_REF> before genotyping; remap PL/AD onto remaining alleles. Estimate cohort allele frequencies via Dirichlet / HWE EM over all sample GLs ([crate::af_calc]), derive site QUAL = −10·log10 P(no variant), assign per-sample genotypes from PL + AF prior (preferring PLs), emit AC/AN/AF/NS/DP/QD. Default emit threshold: standard_confidence_for_calling = 30.0 (hom-ref blocks omitted). Annotation reuse: core INFO via crate::genotyping::compute_core_variant_annotations; QD = QUAL / variant depth. Read-based FS/MQ require pileups and are not computed here.

genotype_site

Genotype-site orchestration surface (L7-B3). Input bundle for region genotyping; production calls assign_genotype_likelihoods_for_region directly.

genotyping

Genotyping implementation

genotyping_dump

Genotyping parity dumps (L2 parity).

given_alleles

GATK AssemblyBasedCallerUtils.addGivenAlleles — forced alleles into assembly. CLI: -alleles is not exposed on gatk-rs HaplotypeCaller (Sprint G). Library wiring exists for L2 c5-force via internal GatkConfig parameters only. See docs/ARCHITECTURE.md (T3-5).

gvcf_writer

GATK GVCFWriter + GATKVCFHeaderLines slice.

gvcf_writer_dump

GVCF writer / header parity dumps.

haplotype

GATK Haplotype representation for assembly output.

haplotype_cigar

GATK CigarUtils.calculateCigar + AlignmentUtils trim/left-align for haplotype realignment.

hc_allele_mapping

GATK AssemblyBasedCallerUtils.createAlleleMapper + spanning-del handling (P1-10).

hc_emit_policy

Shared HC variant emit gates for production strict_java. Combines Java-style AF / confidence checks with optional ASM-8 production admission ([crate::read_event_discovery::is_strict_java_production_emit_admits] — band/motif predicates, not the p12_java_only.tsv oracle; Sprint L-3). Harness-only baseline oracle blocking goes through p12_baseline_emit_oracle_blocks (env + parity_harness). See docs/CLAIM_MATRIX.md.

hc_genotyping_engine

GATK HaplotypeCallerGenotypingEngine slice — production genotyping for strict_java.

hq_soft_clip

HQ soft-clip statistics for HC activity (AVERAGE_HQ_SOFTCLIPS_HQ_BASES_THRESHOLD path).

hq_soft_clip_dump

HQ soft-clip mean dump. Emits the same RCM running mean as ReferenceConfidenceModel#calcGenotypeLikelihoodsOfRefVsAny, not a read-level heuristic.

j_modes_dump

HC output mode emission decisions.

junction_kbest

GATK JunctionTreeKBestHaplotypeFinder / JTBestHaplotype parity.

junction_tree_graph

GATK JunctionTreeLinkedDeBruijnGraph parity: graph build + generateJunctionTrees.

kbest_haplotype

GATK GraphBasedKBestHaplotypeFinder / KBestHaplotype parity (Yen’s-style K-shortest paths on the assembly graph).

likelihood_engine

HC read×haplotype likelihood scoring (, F.4, F.5, F.6). Free functions + HcLikelihoodImplementation enum. PairHMM kernel selection is crate::pairhmm_simd::PairHmmImpl (Log10 default until SIMD gates close).

likelihood_engine_dump

Parity dumps for likelihood engine modes (F.4–F.6).

locus_iterator

GATK IntervalLocusIterator + LocusIteratorByState-style pileup walk (GAP-B-02). Walks every 1-based position in merged intervals (including empty loci) and builds pileups from a monotonically advancing active-read set in BAM iteration order.

locus_pileup_detail_dump

Per-read pileup element dump for RCM / isActive parity (ASM act-2).

minimal_genotyping

GATK MinimalGenotypingEngine path for HC isActive (GAP-C-01). Java delegates to AlleleFrequencyCalculator#calculateSingleSampleBiallelicNonRefPosterior with returnZeroIfRefIsMax = true on PL-rounded likelihoods from uncapped RCM output. Hom-ref capping (getGenotypeLikelihoodsCappedByHomRefLikelihood) is for gVCF emission only, not isActive.

multiallelic_emit

Emit-time multi-allelic coalesce (L7-A2). Production genotyping remains biallelic per ALT; outside contig 2 / chr2, same-POS VCF rows that share a REF and each carry one ALT are merged (ALT=a,b, GT 1/2).

p12_java_format_fixup

P12 L4 harness: apply Java FORMAT/QUAL/AF from parity/fixtures/p12-java-format/all_sites.tsv. Opt-in only: P12_L4_JAVA_FORMAT=1 with --features parity_harness (not implied by P12_PHASE_E).

pairhmm

PairHMM likelihood scaffold. This module provides a deterministic log10-space PairHMM with: explicit Match/Insertion/Deletion state transitions, base + mapping quality integration via effective quality capping, numerically stable log-sum transitions.

pairhmm_dump

PairHMM likelihood dumps for L2 parity.

pairhmm_f3_dump

Parity dumps: BQ caps and haplotype support filtering.

pairhmm_log10

GATK Log10PairHMM (exact log10) — parity with HcParityNativePairHmm / GAP-F-02. R4-1: flat DP matrices + thread-local scratch (same numerics, far fewer allocations).

pairhmm_logless

GATK LoglessPairHMM — linear-space DP, log10 only on the final sum. Observable Java contract (LoglessPairHMM.subComputeReadLikelihoodGivenHaplotypeLog10): INITIAL_CONDITION = 2^1020; free leading deletions = INITIAL / haplen. Match/Ins/Del updates multiply (not log-sum); priors use qualToProb / qualToErrorProb/3. Result = log10(sum last-row M+I) − log10(INITIAL_CONDITION). This is the scalar reference for SIMD kernels. Exact Log10PairHMM remains in crate::pairhmm_log10 for F.1/F.2 bit-identical dumps.

pairhmm_native_dump

GATK Log10PairHMM (GAP-F-02).

pairhmm_qual

GATK PairHMMLikelihoodCalculationEngine base-quality capping.

pairhmm_simd

SIMD / pairs-in-lanes Logless PairHMM (runtime feature dispatch). Parallelization axis: independent haplotypes in SIMD lanes for one shared read (Endeavor-style), not GKL anti-diagonal striping. Scalar crate::pairhmm_logless remains the correctness reference.

parity_harness

Parity harness flags — opt-in env toggles for L3/L4 sign-off only. Production strict_java builds must not set these. Integration tests that need harness behavior must compile with --features parity_harness. See docs/ARCHITECTURE.md (harness env flags).

parity_region_genotype

Java HcParityRegionGenotype + HcParityNativePairHmm path for L2 parity G.2 live gates. Uses raw shard reads (sorted by qname, then start), parity PairHMM defaults (45/45/10), no finalize/overlap correction and no MQ/BQ capping — matching HcFullParityGateDump.assemblyRegionGenotypeSubset.

pcr_error_model

GATK PairHMMLikelihoodCalculationEngine PCR indel error model.

pileup_detection

GATK PileupDetectionArgumentCollection — HC default path + parity supplement. Standard Illumina HaplotypeCaller runs with usePileupDetection=false. When false, Java does not call processPileupAlleles; assembly variation comes from haplotype CIGARs only. Rust parity bridge: [supplement_assembly_pileup_events_from_reads] (strict read pileup) after trim when PileupDetectionConfig::enable_event_supplement is set.

pileup_element

GATK PileupElement flank flags for ReferenceConfidenceModel.isAltBeforeAssembly.

ploidy_dump

Sample ploidy vs activity-evaluation ploidy (L2 parity c5-ploidy).

pre_dragstr_dump

PRE-D01 — DRAGSTR calibration parity scaffold.

read_assembly_filter

GATK HaplotypeCallerEngine.filterNonPassingReads (assembly / genotyping path).

read_downsample

GATK PositionalDownsampler — cap reads per alignment start. Reservoir mode matches GATK ReservoirDownsampler + Utils.getRandomGenerator (seed 47382911).

read_filter_dump

HC read-filter decision dump.

read_header_semantics

SAM/BAM header semantics (RG/PG) validation for user inputs.

read_model

Read model semantics aligned with GATK HaplotypeCallerEngine read filters. Java reference: HaplotypeCallerEngine.makeStandardHCReadFilters and WellformedReadFilter (CigarUtils.isGood, ReadFilterLibrary, ReadUtils.alignmentAgreesWithHeader).

read_pre_len_dump

PRE.2 gate dump: read-pre-len.

read_pre_mq_dump

PRE.3 gate dump: read-pre-mq.

read_pre_overlap_dump

PRE.4 gate dump: read-pre-overlap.

read_projection

Read ↔ reference projection and clipping helpers. Coordinates follow BAM conventions: alignment_start is 0-based inclusive first aligned reference position; query_index is 0-based into the read sequence as stored in the BAM record (soft-clipped bases are present; hard-clipped bases are absent).

read_realignment

GATK AssemblyBasedCallerUtils.realignReadsToTheirBestHaplotype (P1⁴ REALIGN).

read_threading_assembler

GATK ReadThreadingAssembler production path on crate::assembly::AssemblyGraph.

read_threading_graph

GATK ReadThreadingGraph / AbstractReadThreadingGraph parity builder. Replaces naive de Bruijn per-read edge counting with read threading: unique-kmer starts, backward multiplicity on incoming matches, forward extension by suffix, and per-sample pruning multiplicity (MultiSampleEdge with numPruningSamples = 1).

read_transformer

GATK ReadTransformer hooks for shard read preparation. Java MultiIntervalLocalReadShard order: pre-transform → read filter → post-transform → downsampler.

read_transformer_dump

HC shard read-pipeline dump: IUPAC pre-transform, filter, post-transform. Downsample is not included (isolated in D.2). Matches HcFullParityGateDump read-shard-pipeline.

read_unclip

GATK read soft-clip preparation before assembly. Mirrors AssemblyBasedCallerUtils.finalizeRegion soft-clip branch and ReadClipper / CigarUtils.revertSoftClips.

read_unclip_dump

PRE.1 gate dump: read-pre-softclip.

read_validation

Mapped-read sanity checks for user BAM fields (lengths / coordinates).

ref_confidence

GATK ReferenceConfidenceModel slice — ref-vs-any GLs and inactive-region modeling. Used for gVCF / no-variation loci. Dense-band RCM reconciliation that applies P12 gradation tables is interval-scoped (waiver W-H3) in crate::reference_vcf_emit not a genome-wide claim. See docs/CLAIM_MATRIX.md.

ref_confidence_dump

ReferenceConfidenceModel parity dumps.

ref_confidence_merger

GATK ReferenceConfidenceVariantContextMerger (H-D02) — diploid gVCF merge slice.

ref_confidence_merger_dump

H-D02 — ReferenceConfidenceVariantContextMerger parity dumps.

reference_context

GATK ReferenceContext for assembly-region apply / callRegion. Wraps a 1-based inclusive reference interval and lazily-loaded bases from FASTA, matching new ReferenceContext(reference, assemblyRegion.getPaddedSpan) in Java.

reference_vcf_emit

Inactive-region and gVCF reference-confidence VCF emission.

region_read_likelihood

Leaf likelihood-row type (breaks engine ↔ genotyping module cycle). Read×haplotype likelihood matrix cell (PairHMM / callRegion). Leaf domain type shared by the engine and genotyping — kept out of crate::engine so genotyping does not depend on the full callRegion engine module.

region_vcf_dump

VCF record parity dumps (j2-vcf, j2-format).

region_vcf_emit

Emit VCF variant records from a call_region outcome (production path). Applies crate::hc_emit_policy gates, optional HC site annotations, and (harness-only) Java FORMAT overlay via crate::p12_java_format_fixup. Default CLI emission does not require oracle TSV files. See docs/ARCHITECTURE.md stage 11.

run

HaplotypeCaller CLI pipeline entry. gVCF RCM scope (H1-3): post-walk reconcile_p12_cluster_rcm_band reconcile_fragmented_dense_cluster_bands run only when walker layout fragments P12 dense cluster bands on chr2:92300000–92350000. Not genome-wide Java RCM parity.

runtime_config

Process env → typed config (Sprint I-4). Prefer over scattered std::env::var. Central runtime configuration parsed from environment variables (Sprint I-4). Production algorithm modules must not call std::env::var directly — use this module or crate::parity_harness (harness-gated flags only). Inventory: docs/ARCHITECTURE.md.

semantic_trace

Optional observe-only semantic checkpoints (GATK_RS_SEMANTIC_TRACE). Optional structured semantic checkpoints for HaplotypeCaller (observe-only). Enable with GATK_RS_SEMANTIC_TRACE=<ndjson-path>. When unset, every emit helper is a single atomic load and returns immediately — no algorithm branching. Schema id: gatk_rs.hc.semantic_trace/v1. Compare Java vs Rust with scripts/parity/compare_semantic_trace.py.

seq_graph

GATK SeqGraph conversion and cleanup (BaseGraph.toSequenceGraph, SeqGraph.cleanup).

seq_kbest_haplotype

GATK GraphBasedKBestHaplotypeFinder on SeqGraph (Path.getBases stitching).

shared_bam

Arc-backed BAM records shared between the shard read cache and assembly regions. Shared BAM records for shard/region ownership without deep clones.

smith_waterman

GATK SmithWatermanJavaAligner parity (JAVA implementation path). Score / backtrack grids are flat row-major Vec<i32> (one allocation each) instead Vec<Vec<i32>> (one allocation per row).

variant_site_hc_annotations

GATK HC variant-site INFO / QUAL for assembly-region VCF emission (J-D01).

walker

GATK4 AssemblyRegionWalker–style read sharding. See docs/ARCHITECTURE.md. Mirrors the first step of AssemblyRegionWalker.makeReadShards: group user intervals by contig, merge overlaps on the input grid, then expand each span by assemblyRegionPadding (default 100 bp per side, matching AssemblyRegionArgumentCollection.DEFAULT_ASSEMBLY_REGION_PADDING in GATK 4.x). B.2: region stream over each padded span lives in crate::assembly_region_iterator (AssemblyRegionIterator). B.3: apply vs callRegion disposition — crate::walker_apply and crate::HaplotypeCallerEngine::walker_apply_stats.

walker_apply

AssemblyRegionWalker.apply vs HaplotypeCallerEngine.callRegion disposition. Java: AssemblyRegionWalker invokes apply once per region from AssemblyRegionIterator. HaplotypeCallerEngine.callRegion takes the inactive fast path (no graph / local assembly) when !region.isActive, and otherwise runs assembly + genotyping.

walker_traversal

GATK4 AssemblyRegionWalker#traverse orchestration. Chains B.1 read shards → B.2 AssemblyRegionIterator → B.3 WalkerApplyStats.

worker_threads

Rayon global pool / RAYON_NUM_THREADS initialization for HaplotypeCaller.

Structs

AnnotatedSite

Annotated site after parity v1 core INFO computation.

BamoutWriter

Placeholder BAM writer counting assembled-read emissions for parity gates.

BamoutWriterConfig

Enable flag for assembled-read BAM output (product stub).

RegionPileupLocus

One locus pileup retained on an crate::assembly_region_iterator::AssemblyRegion when pileup tracking is enabled (GATK AlignmentAndReferenceContext).

VariantAnnotationContext

Input for parity v1 site annotation.

Well19937c

org.apache.commons.math3.random.Well19937c (WELL19937c, K=19937).

Enums

MappingQuality

SAM/BAM mapping quality, distinguishing unavailable (255) from a real score.

OptionalTagValue

Typed SAM optional field value (parity subset: i, f, Z, A).

Constants

GATK_WELL19937C_SEED

GATK Utils.GATK_RANDOM_SEED — also seeds RandomDataGenerator(Well19937c).

PARITY_V1_FORMAT_KEYS

FORMAT keys agreed for parity v1 (computed in; listed for manifest completeness).

PARITY_V1_INFO_KEYS

INFO keys emitted by parity v1 (i1-core gate).

Functions

annotate_parity_v1_site

Compute parity v1 core INFO (AC/AN/AF/NS/DP) for one site.

apply_contamination_to_pileup

Apply contamination downsampling to a single-sample pileup (in-place).

count_reads_overlapping_tile

Deterministic ingress order helpers (foundation Step 35 + HC tiling). Count primary alignments whose reference span overlaps tile (linear BAM scan).

dump_allele_biased_evidence_locus_tsv

Dump evidence removal at one locus: removed_count, sorted removed_qnames.

dump_bamout_stub_tsv

J-D04 — stable bamout gate dump.

dump_target_allele_counts_tsv

Dump targetAlleleCounts parity row (/ gate d2c).

filtered_read_iteration_order

Deterministic ingress order helpers (foundation Step 35 + HC tiling). Deterministic filtered ingress order for reads with aligned reference span. The returned order follows file iteration order from HTSlib and is used as a contract for ingress reproducibility checks.

format_optional_tag_field

parse_optional_tag_field

select_allele_biased_evidence_indices

Indices into evidence lists to remove (GATK selectAlleleBiasedEvidence).

strict_java_asm8_only_enabled

Production ASM-8 / P12-bridge env flags (public so integration tests need no parity_harness). Graph-only strict production (CIGAR/EventMap finalize, java-only emit pin). Production default: ASM-8 when P12 bridges are off (no env required). GATK_RS_ASM8_ONLY=1: harness-only explicit toggle (see HARNESS_FLAGS.md).

strict_java_p12_ensure_bridges_enabled

Production ASM-8 / P12-bridge env flags (public so integration tests need no parity_harness). P12 list-inject / post-HMM ensure_* bridges. Opt-in: GATK_RS_P12_ENSURE_BRIDGES=1 (harness only). Forced off when GATK_RS_ASM8_ONLY=1 (harness).

target_allele_counts

GATK AlleleBiasedDownsamplingUtils.targetAlleleCounts.

total_read_tile_overlaps

Deterministic ingress order helpers (foundation Step 35 + HC tiling). Sum count_reads_overlapping_tile across all tiles (read may be counted in multiple tiles).